+++ ClickerGame/ResourceManagement/ResourceAdjustmentFactory.cs
@@ -2,10 +2,10 @@
 {
     public class ResourceAdjustmentFactory
     {
-        public IResourceAdjustment CreateIncreaseEqualTo(double adjustment) =>
+        public IResourceAdjustment CreateIncreaseEqualTo(decimal adjustment) =>
             new ResourceAdjustment(adjustment);
 
-        public IResourceAdjustment CreateDecreaseEqualTo(double adjustment) =>
+        public IResourceAdjustment CreateDecreaseEqualTo(decimal adjustment) =>
             new ResourceAdjustment(adjustment * -1);
     }
 }
